Fixed a bug
I had an error at startup along with NVDA load.
If you click reload add-ons in the Tools menu, there was no error anymore.
I looked at the log and realized what the problem was.
Fix in the code.

P.S. It's a bad idea to declare variables inside a class. This should be done through the init  constructor.